+++ Assets/_SOURCE/Gameplay/Characters/Pets/Hens/_components/HenMover.cs
@@ -7,7 +7,7 @@ namespace Gameplay.Characters.Pets.Hens
   {
     [Inject] private CharacterController _characterController;
 
-    public void Move(Vector3 target, int moveSpeed)
+    public void Move(Vector3 target, float moveSpeed)
     {
       Vector3 direction = (target - transform.position).normalized;
       _characterController.Move(direction * (moveSpeed * Time.deltaTime));
