+++ Assets/_SOURCE/_Infrastructure/StaticDataServices/IStaticDataService.cs
@@ -24,7 +24,7 @@ namespace StaticDataServices
     LootConfig GetLootConfig(CurrencyId lootDropId);
     MusicConfig GetMusicConfig(MusicId musicId);
     SoundConfig GetSoundConfig(SoundId soundId);
-    int GetInitialStat(StatId id);
+    float GetInitialStat(StatId id);
     Dictionary<QuestId, QuestConfig> GetQuestConfigs();
     QuestConfig GetQuestConfig(QuestId questId);
     ExpirienceConfig GetExpirienceConfig();
